requirements.txt
PyYAML==5.1.2
records==0.5.3
loguru==0.3.2
requests==2.22.0
beautifulsoup4==4.8.1
Scrapy==1.8.0